requirements.txt
# Основные зависимости для HR-Психоаналитического бота v2.0
python-telegram-bot[job-queue]>=20.0
openai>=1.0.0
python-dotenv>=1.0.0

# ИИ и анализ данных
textblob>=0.17.1
vaderSentiment>=3.3.2
scikit-learn>=1.3.0
numpy>=1.24.0
tiktoken>=0.5.0

# Валидация и конфигурация
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Логирование и мониторинг
structlog>=23.0.0
prometheus-client>=0.17.0

# Тестирование
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Дополнительные зависимости
requests>=2.31.0
asyncio-mqtt>=0.13.0
pyyaml>=6.0